RecipeSideServer/videoServer/fastApiProject/requirements.txt
adbc_driver_manager==1.3.0
adbc_driver_postgresql==1.3.0
adbc_driver_sqlite==1.3.0
aenum==3.1.15
aesara==2.9.4
aiohttp==3.11.0
amdsmi==6.2.4
AppKit==0.2.8
astunparse==1.6.3
attr==0.3.2
beautifulsoup4==4.12.3
boto3==1.35.23
botocore==1.35.23
cashew==0.4.1
charset_normalizer==2.0.4
conda==4.3.16
ConfigParser==7.1.0
contextlib2==21.6.0
coremltools==8.0
cryptography==42.0.2
ctypes_snappy==1.2
cycler==0.11.0
Cython==3.0.11
#decord==0.6.0
#deeplearning==0.0.1x
defusedxml==0.7.1
dill==0.3.7
disco==1.40.4
dl==0.1.0
docutils==0.18.1
docutils==0.18.1
etcd==2.0.8
expecttest==0.2.1
fastapi==0.115.5
fastparquet==2024.11.0
fbscribelogger==0.1.7
flint==0.0.1
Foundation==0.1.0a0.dev1
gdown==5.2.0
gmpy2==2.1.2
h5py==3.12.1
hickle==5.0.3
HTMLParser==0.0.2
HTMLParser==0.0.2
hypothesis==6.118.8
input==0.0.0
ipython==8.12.3
ipywidgets==7.6.5
isal==1.7.1
jnius==1.1.0
JPype1==1.5.0
kerchunk==0.2.6
keyring==23.13.1
lmdb==1.4.1
lxml==5.3.0
lz4==4.3.2
lzmaffi==0.3.0
matchpy==0.5.5
matplotlib==3.8.0
mock==5.1.0
MonkeyType==23.3.0
moviepy==1.0.3
mtrand==0.1
multiset==2.1.1
mypy==1.8.0
nose==1.3.7
numarray==1.5.1
Numeric==24.2
numexpr==2.8.7
odfpy==1.4.1
olefile==0.47
onnxscript==0.1.0.dev20241112
openai==0.28.0
openpyxl==3.0.10
optree==0.13.1
paramiko==3.5.0
pickle5==0.0.12
protobuf==4.21.12
psutil==5.9.0
pycocotools==2.0.8
pycosat==0.6.6
pydantic==1.10.12
pydot==3.0.2
pydy==0.7.1
pygit2==1.16.0
pyglet==2.0.18
pygraphviz==1.14
pyinstaller==6.11.1
pyinstrument==5.0.0
pymc==5.18.0
pymc3==3.11.6
pynvml==11.5.3
pyobjc_framework_Cocoa==9.0
pyOpenSSL==24.2.1
pyOpenSSL==24.2.1
PyQt4==4.11.4
PyQt5==5.15.11
PyQt5_sip==12.13.0
PyQt6==6.7.1
PyQt6_sip==13.8.0
pysat==3.2.1
PySide6==6.8.0.2
PySide6==6.8.0.2
PySide6_Addons==6.8.0.2
PySide6_Essentials==6.8.0.2
pytest==7.4.0
python-dotenv==1.0.1
python_calamine==0.3.1
pytorch_lightning==2.4.0
pyxlsb==1.0.10
QtPy==2.4.1
railroad==0.5.0
requests_kerberos==0.15.0
s3fs==0.6.0
safetensors==0.4.5
sage==0.0.0
scikit_learn==1.2.2
scipy==1.14.1
scipy_doctest==1.4
setuptools_scm==8.1.0
Shapely==2.0.6
simplejson==3.19.3
slack_sdk==3.33.3
smbprotocol==1.15.0
Sphinx==5.0.2
SQLAlchemy==2.0.25
stream==0.8
symengine==0.13.0
tables==3.9.2
Theano==1.0.5
thread==2.0.5
threadpoolctl==3.5.0
torch_xla==2.5.1
torcharrow==0.1.0
torchaudio==2.5.1
torchrec==1.0.0
tornado==6.3.3
traitlets==5.14.3
trove_classifiers==2024.10.21.16
ujson==5.4.0
urllib3_secure_extra==0.1.0
uvicorn==0.32.0
wcwidth==0.2.5
xarray==2023.6.0
xlrd==2.0.1
XlsxWriter==3.2.0
xmlrpclib==1.0.1
xmlrunner==1.7.7
yarl==1.17.1
z3==0.2.0
zstandard==0.19.0
